Allow no-restricted-paths to restrict from packages
I am working in a project where we have a package installed that has
a "private" directory with some files in it that are not intended to be
part of the public API. It would be nice to be able to restrict these
files from being imported with configuration like:

```json
"import/no-restricted-paths": ["error", {
  "zones": [
    { "target": ".", "from": "my-package/private" },
  ],
}],
```

This commit takes a swing at making this work. I am not entirely sure if
this is the right approach and I was unable to get the specs running the
way I wanted, but I hope this can at least get the ball rolling.