Add test case for import/resolve extensions order
When there are multiple files with same names and different extensions,
the resolver will go by extensions in the order specified in
import/resolve settings.

This test case ensures that the "extensions" rule is correctly forcing
the extension when omitting it would cause it to resolve to the wrong
file.

Closes #965.