Do not report type imports from flow-typed packages
When importing types from packages that are defined in a file inside
the `flow-typed` directory, prevent the `named` rule from reporting
that the package itself does not provide an export with that name.

Currently there is support for type imports from local files. Restrict
these checks to imports with relative paths.