[Fix] `named`/`ExportMap`: Fix ExportMap for a merged typescript namespace
A typescript namespace may be declared multiple times, it is then merged
together and considered to be a single declaration from consuming code.
In the case where a merged namespace is assigned as the export from a
module then the declarations from all the instances of the namespace in
the AST need to be considered as exported.