[utils] [patch] mark eslint as an optional peer dep
This allows it to find eslint from module-require.js when the package manager is strict.

Maintainer note: this should not be necessary, but using `peerDependenciesMeta` here seems to have no downsides for real npm users, and allows tools like pnpm or yarn PnP to statically "know" that this package wants to access `eslint` if it is present.