[Fix] `no-unused-modules`: don't error out when running with flat config and an eslintrc isn't present
This change adjusts how we're instantiating the FileEnumerator from eslint's unsupported api, in the case that the user is running with flat config.  We have to turn off the `useEslintrc` property on the ConfigArrayFactory that's passed into the FileEnumerator's constructor.

Note: This doesn't fix the fact that the FileEnumerator doesn't have knowledge of what the user's config is ignoring, it just prevents the rule from looking for a legacy / rc config and erroring out.  FileEnumerator used the rc config to understand which files to ignore.